refactor: WebUIChessApplication 에서 라우팅 이외의 부분 Controller 로 분리
